feat(book): implement book management module
- Add book-card and rating components
- Implement book-list, my-books, and manage-book pages
- Optimize manage-book form layout for laptop screens
- Integrate backend validation error handling

+++ book-network/src/main/java/com/yladraoui/book/common/BaseEntity.java
@@ -41,5 +41,5 @@ public class BaseEntity {
 
     @LastModifiedBy
     @Column(insertable = false)
-    private Integer lastModifiedBy;
+    private Long lastModifiedBy;
 }
